:books: p23

+++ part_01/SimpleAdder/SimpleAdder.cpp
@@ -2,15 +2,15 @@
 
 int main()
 {
-	int val1;
+	double val1;
 	std::cout << "첫 번째 숫자 입력 = ";
 	std::cin >> val1;
 
-	int val2;
+	double val2;
 	std::cout << "두 번째 숫자 입력 = ";
 	std::cin >> val2;
 
-	int result = val1 + val2;
+	double result = val1 + val2;
 	std::cout << "덧셈 결과 = " << result << std::endl;
 	return 0;
 }
